update tests for FlorisModel get_param, set_param, print_dict, nested_get, and nested_set
